requirements.txt
flask
transformers
torch
ultralytics
pillow
